requirements.txt
tensorflow
mlflow
zenml
deepchecks
pillow
streamlit
scikit-learn
pandas
numpy
opencv-python
